handle resize of window, more readable code and fix requests from codereview
